updated error handling
