+++ common4j/src/main/com/microsoft/identity/common/java/commands/WebAppsTokenResponse.kt
@@ -20,11 +20,38 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.broker.ipc
+package com.microsoft.identity.common.java.commands
+
+import com.google.gson.annotations.SerializedName
 
 /**
- * Error response for WebApps token requests.
+ * Successful response for WebApps token requests.
  */
-data class WebAppsTokenErrorResponse(
-    val throwable: Throwable? = null
-)
+data class WebAppsTokenResponse(
+    @SerializedName(FIELD_USER_NAME)
+    val userName: String,
+
+    @SerializedName(FIELD_HOME_ACCOUNT_ID)
+    val homeAccountId: String,
+
+    @SerializedName(FIELD_EXPIRES_IN)
+    val expiresIn: String,
+
+    @SerializedName(FIELD_ID_TOKEN)
+    val idToken: String,
+
+    @SerializedName(FIELD_ACCESS_TOKEN)
+    val accessToken: String,
+
+    @SerializedName(FIELD_PROPERTIES)
+    val properties: Map<String, String>? = null
+) {
+    companion object {
+        const val FIELD_USER_NAME = "userName"
+        const val FIELD_HOME_ACCOUNT_ID = "homeAccountId"
+        const val FIELD_EXPIRES_IN = "expiresIn"
+        const val FIELD_ID_TOKEN = "idToken"
+        const val FIELD_ACCESS_TOKEN = "accessToken"
+        const val FIELD_PROPERTIES = "properties"
+    }
+}
